feat: Integrate bundle analyzer and refactor login components to use LoginWrapper
