+++ src/app/me/login/google/page.tsx
@@ -1,32 +1,14 @@
 'use client';
 
-import { useRouter, useSearchParams } from 'next/navigation';
-import { useEffect, useState } from 'react';
+import { useSearchParams } from 'next/navigation';
 import { loginGoogle } from '@/lib/apiManager';
-import { httpStatusCodes } from '@/lib/StatusCodes';
-import LoadingWrapper from '@/components/me/accounts/LoadingWrapper';
+import LoginWrapper from '@/components/me/accounts/LoginWrapper';
 
 const Page = () => {
-    const router = useRouter();
     const params = useSearchParams();
     const code = params.get('code');
-    const [loadingStatus, setLoadingStatus] = useState<string>('');
 
-    useEffect(() => {
-        if (!code) router.replace('/me');
-
-        loginGoogle(code)
-            .then(() => router.replace('/me'))
-            .catch(response => {
-                setLoadingStatus(
-                    `${response.status}: ${
-                        response.data.message || httpStatusCodes[response.status]
-                    }`
-                );
-            });
-    }, []);
-
-    return <LoadingWrapper loadingStatus={loadingStatus} />;
+    return <LoginWrapper code={code} callback={loginGoogle} />;
 };
 
 export default Page;
